Random kapetz

+++ stack/murmurhash.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-unsigned int murmurHash2 (const void* key, size_t len);
+unsigned int murmurHash2 (const void* key, unsigned int len);
 
 #endif
